#17 Moved the ScopeElement association to Scope.java to match CML files.
